Organize API contracts by purpose
