+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/events/ServerHoldChangedEvent.java
@@ -1,6 +1,6 @@
 package com.akselglyholt.velocitylimbohandler.api.events;
 
-import com.akselglyholt.velocitylimbohandler.api.HoldSnapshot;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldSnapshot;
 
 import java.util.List;
 import java.util.Objects;
