+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/hold/HoldResult.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.hold;
 
 import java.util.Objects;
 import java.util.Optional;
