+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/hold/HoldSnapshot.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.hold;
 
 import java.time.Instant;
 import java.util.Objects;
