+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/hold/HoldTarget.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.hold;
 
 /** Type of resource protected by a hold lease. */
 public enum HoldTarget {
