+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/lifecycle/Availability.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.lifecycle;
 
 /** Runtime availability of Velocity Limbo Handler. */
 public enum Availability {
