+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/queue/QueueTier.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.queue;
 
 /** Read-only, permission-derived queue tier. */
 public enum QueueTier {
