+++ api/src/main/java/com/akselglyholt/velocitylimbohandler/api/queue/QueuedPlayerSnapshot.java
@@ -1,4 +1,4 @@
-package com.akselglyholt.velocitylimbohandler.api;
+package com.akselglyholt.velocitylimbohandler.api.queue;
 
 import java.util.Objects;
 import java.util.UUID;
