+++ src/main/java/com/akselglyholt/velocityLimboHandler/api/VelocityLimboApiImpl.java
@@ -2,32 +2,32 @@ package com.akselglyholt.velocityLimboHandler.api;
 
 import com.akselglyholt.velocityLimboHandler.VelocityLimboHandler;
 import com.akselglyholt.velocityLimboHandler.storage.PlayerManager;
-import com.akselglyholt.velocitylimbohandler.api.Availability;
-import com.akselglyholt.velocitylimbohandler.api.EnterRequest;
-import com.akselglyholt.velocitylimbohandler.api.EnterResult;
-import com.akselglyholt.velocitylimbohandler.api.HoldLease;
-import com.akselglyholt.velocitylimbohandler.api.HoldReleaseResult;
-import com.akselglyholt.velocitylimbohandler.api.HoldRequest;
-import com.akselglyholt.velocitylimbohandler.api.HoldResult;
-import com.akselglyholt.velocitylimbohandler.api.HoldSnapshot;
-import com.akselglyholt.velocitylimbohandler.api.HoldStatus;
-import com.akselglyholt.velocitylimbohandler.api.HoldTarget;
 import com.akselglyholt.velocitylimbohandler.api.LimboController;
-import com.akselglyholt.velocitylimbohandler.api.LimboPhase;
-import com.akselglyholt.velocitylimbohandler.api.ManagedPlayerSnapshot;
-import com.akselglyholt.velocitylimbohandler.api.QueueSnapshot;
-import com.akselglyholt.velocitylimbohandler.api.QueueSummary;
-import com.akselglyholt.velocitylimbohandler.api.QueueTier;
-import com.akselglyholt.velocitylimbohandler.api.QueuedPlayerSnapshot;
-import com.akselglyholt.velocitylimbohandler.api.ReconnectOutcome;
-import com.akselglyholt.velocitylimbohandler.api.RetargetResult;
 import com.akselglyholt.velocitylimbohandler.api.VelocityLimboApi;
+import com.akselglyholt.velocitylimbohandler.api.entry.EnterRequest;
+import com.akselglyholt.velocitylimbohandler.api.entry.EnterResult;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerEnteredLimboEvent;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerLeftLimboEvent;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerLimboStateChangedEvent;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerReconnectAttemptEvent;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerReconnectResultEvent;
 import com.akselglyholt.velocitylimbohandler.api.events.ServerHoldChangedEvent;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldLease;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldReleaseResult;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldRequest;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldResult;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldSnapshot;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldStatus;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldTarget;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.Availability;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.LimboPhase;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.ReconnectOutcome;
+import com.akselglyholt.velocitylimbohandler.api.player.ManagedPlayerSnapshot;
+import com.akselglyholt.velocitylimbohandler.api.player.RetargetResult;
+import com.akselglyholt.velocitylimbohandler.api.queue.QueueSnapshot;
+import com.akselglyholt.velocitylimbohandler.api.queue.QueueSummary;
+import com.akselglyholt.velocitylimbohandler.api.queue.QueueTier;
+import com.akselglyholt.velocitylimbohandler.api.queue.QueuedPlayerSnapshot;
 import com.velocitypowered.api.plugin.PluginContainer;
 import com.velocitypowered.api.proxy.ConnectionRequestBuilder;
 import com.velocitypowered.api.proxy.Player;
