+++ src/main/java/com/akselglyholt/velocityLimboHandler/storage/PlayerManager.java
@@ -4,8 +4,8 @@ import com.akselglyholt.velocityLimboHandler.VelocityLimboHandler;
 import com.akselglyholt.velocityLimboHandler.api.VelocityLimboApiImpl;
 import com.akselglyholt.velocityLimboHandler.misc.MessageFormatter;
 import com.akselglyholt.velocityLimboHandler.misc.Utility;
-import com.akselglyholt.velocitylimbohandler.api.Availability;
-import com.akselglyholt.velocitylimbohandler.api.ReconnectOutcome;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.Availability;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.ReconnectOutcome;
 import com.velocitypowered.api.proxy.Player;
 import com.velocitypowered.api.proxy.server.RegisteredServer;
 import dev.dejvokep.boostedyaml.route.Route;
