+++ src/test/java/com/akselglyholt/velocityLimboHandler/api/VelocityLimboApiImplTest.java
@@ -2,15 +2,15 @@ package com.akselglyholt.velocityLimboHandler.api;
 
 import com.akselglyholt.velocityLimboHandler.VelocityLimboHandler;
 import com.akselglyholt.velocityLimboHandler.storage.PlayerManager;
-import com.akselglyholt.velocitylimbohandler.api.HoldReleaseResult;
-import com.akselglyholt.velocitylimbohandler.api.HoldRequest;
-import com.akselglyholt.velocitylimbohandler.api.HoldStatus;
-import com.akselglyholt.velocitylimbohandler.api.EnterRequest;
-import com.akselglyholt.velocitylimbohandler.api.EnterResult;
 import com.akselglyholt.velocitylimbohandler.api.LimboController;
-import com.akselglyholt.velocitylimbohandler.api.LimboPhase;
-import com.akselglyholt.velocitylimbohandler.api.RetargetResult;
+import com.akselglyholt.velocitylimbohandler.api.entry.EnterRequest;
+import com.akselglyholt.velocitylimbohandler.api.entry.EnterResult;
 import com.akselglyholt.velocitylimbohandler.api.events.PlayerLeftLimboEvent;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldReleaseResult;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldRequest;
+import com.akselglyholt.velocitylimbohandler.api.hold.HoldStatus;
+import com.akselglyholt.velocitylimbohandler.api.lifecycle.LimboPhase;
+import com.akselglyholt.velocitylimbohandler.api.player.RetargetResult;
 import com.velocitypowered.api.event.EventManager;
 import com.velocitypowered.api.plugin.PluginContainer;
 import com.velocitypowered.api.plugin.PluginDescription;
